Final adjustments and replacement of old PVT
