+++ frontend/src/modules/Pvt/components/DependentVariableSelector/index.ts
@@ -1,0 +1,2 @@
+export { DependentVariableSelector } from "./dependentVariableSelector";
+export type { DependentVariableSelectorProps } from "./dependentVariableSelector";
