--- frontend/src/modules/PvtRefactored/preview.tsx
@@ -1,6 +1,0 @@
-import { DrawPreviewFunc } from "@framework/Preview";
-import previewImg from "./preview.svg";
-
-export const preview: DrawPreviewFunc = function (width: number, height: number) {
-   return <img src={previewImg} style={{width, height}} />
-};
