--- frontend/src/modules/PvtRefactored/registerModule.tsx
@@ -1,10 +1,0 @@
-import { ModuleRegistry } from "@framework/ModuleRegistry";
-
-import { preview } from "./preview";
-import { Interface, State } from "./state";
-
-ModuleRegistry.registerModule<State, Interface>({
-    moduleName: "PvtRefactored",
-    defaultTitle: "PVT (refactored)",
-    preview,
-});
